A quest may ask for more of an item than a uint16 can count
Quest::Quest read RequiredItemCount through Field::GetUInt16, which aborts the
worldserver when the stored count does not round-trip through 16 bits. The
member it is read into and the field the quest packets carry are both uint32,
and only the stock column is narrow, so the getter was the one place that could
not carry a large count - a currency quest priced in hundreds of thousands of a
token brought the realm down at "Loading Quests..." rather than being rejected.

Read it as uint32. A narrow column widens cleanly through the same round-trip
check, so this is correct whether or not the column is ever widened.

+++ src/server/game/Quests/QuestDef.cpp
@@ -110,7 +110,11 @@ Quest::Quest(Field* questRecord)
     for (uint32 i = 0; i < QUEST_ITEM_OBJECTIVES_COUNT; ++i)
     {
         RequiredItemId[i] = questRecord[87+i].GetUInt32();
-        RequiredItemCount[i] = questRecord[93+i].GetUInt16();
+        // Read as uint32: Quest::RequiredItemCount and the field the quest packets
+        // carry are both uint32, and Field::GetUInt16 aborts the server on a count
+        // that does not fit - a currency quest asking for more than 65535 of an
+        // item is a legitimate thing to write, and the stock column is only narrow.
+        RequiredItemCount[i] = questRecord[93+i].GetUInt32();
 
         if (RequiredItemId[i])
             ++_reqItemsCount;
